+++ castle/cms/cron/_social_counts.py
@@ -1,65 +1,89 @@
-# https://gist.github.com/jonathanmoore/2640302
+import json
+import logging
+import time
+from multiprocessing import Pool
+from urllib import urlencode
+
+import requests
+
+import transaction
 from AccessControl.SecurityManagement import newSecurityManager
 from Acquisition import aq_parent
 from BTrees.OOBTree import OOBTree
 from castle.cms.social import COUNT_ANNOTATION_KEY
-from castle.cms.utils import clear_object_cache
-from castle.cms.utils import index_in_es
-from castle.cms.utils import retriable
-from multiprocessing import Pool
+from castle.cms.utils import clear_object_cache, index_in_es, retriable
 from plone.app.redirector.interfaces import IRedirectionStorage
 from plone.registry.interfaces import IRegistry
 from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone import defaultpage
 from Products.CMFPlone.interfaces.siteroot import IPloneSiteRoot
 from tendo import singleton
-from urllib import quote_plus
 from zope.annotation.interfaces import IAnnotations
 from zope.component import getUtility
 from zope.component.hooks import setSite
 
-import json
-import logging
-import requests
-import time
-import transaction
 
+USE_MULTIPROCESSING = True
+
+MATOMO_TOKEN_AUTH = 'castle.matomo_token_auth'
+MATOMO_BASE_URL = 'castle.matomo_base_url'
+MATOMO_SITE_ID = 'castle.matomo_site_id'
 
 logger = logging.getLogger('castle.cms')
 
 
-def get_facebook_url_data(urls):
-    urls_string = '("%s")' % (
-        '","'.join(quote_plus(url) for url in urls)
+def get_matomo_api_url(url):
+    registry = getUtility(IRegistry)
+    site_id = registry.get(MATOMO_SITE_ID, None)
+    base_url = registry.get(MATOMO_BASE_URL, None)
+    token_auth = registry.get(MATOMO_TOKEN_AUTH, None)
+    params = {
+        'module': 'API',
+        'method': 'Actions.getOutlinks',
+        'idSite': site_id,
+        'period': 'year',
+        'date': 'today',
+        'format': 'json',
+        'token_auth': token_auth,
+        'expanded': '1',
+        'filter_pattern_recursive': '',
+        'u': url
+    }
+    return '{}/?{}'.format(
+        base_url,
+        urlencode(params)
     )
-    access_url = COUNT_URLS['facebook']['url'] % urls_string
-    resp = requests.get(access_url).content
-    data = json.loads(resp)
-    count = 0
-    if 'data' not in data:
-        return 0
-    data = data['data']
-    for group in data:
-        count += group.get('like_count', 0)
-        count += group.get('total_count', 0)
-        count += group.get('share_count', 0)
-        count += group.get('comment_count', 0)
-    return count
 
 
-COUNT_URLS = {
-    'pinterist': {
+_matomo_data_mapping = {
+    'facebook_matomo': 'www.facebook.com',
+    'twitter_matomo': 'twitter.com'
+}
+
+
+def get_matomo_url_data(urls):
+    data = {
+        'facebook_matomo': 0,
+        'twitter_matomo': 0
+    }
+    for url in urls:
+        query_url = get_matomo_api_url(url)
+        resp = requests.get(query_url, timeout=10).content
+        datatable = json.loads(resp)
+        for d in datatable:
+            for key, label in _matomo_data_mapping.items():
+                if d['label'] == label:
+                    data[key] += d['nb_visits']
+    return data
+
+
+COUNT_TYPES = {
+    'pinterest': {
         'url': 'http://api.pinterest.com/v1/urls/count.json?callback=foobar&url=%s',
         'slash_matters': True
     },
-    'linkedin': {
-        'url': 'https://www.linkedin.com/countserv/count/share?format=json&url=%s'
-    },
-    'facebook': {
-        'url': 'https://graph.facebook.com/fql?q=SELECT like_count, total_count, '
-               'share_count,click_count,comment_count FROM link_stat '
-               'WHERE url in %s',
-        'generator': get_facebook_url_data
+    'matomo': {
+        'generator': get_matomo_url_data,
     }
 }
 
@@ -68,10 +92,13 @@ def _get_url_data(args):
     config, url = args
     access_url = config['url'] % url
     resp = requests.get(access_url).content.lstrip('foobar(').rstrip(')')
-    data = json.loads(resp)
     try:
+        data = json.loads(resp)
+    except ValueError:
+        return 0
+    if 'count' in data:
         return data['count']
-    except KeyError:
+    else:
         return 0
 
 
@@ -88,41 +115,32 @@ def _get_urls_data(args):
         for orig_url in urls:
             req_urls.append((config, orig_url))
             if config.get('slash_matters'):
-                req_urls.append((config, orig_url + '/'))
-                req_urls.append((config, _swap_protocol(orig_url + '/')))
-        results = _req_pool.map(_get_url_data, req_urls)
+                req_urls.append((config, orig_url.rstrip('/') + '/'))
+        if USE_MULTIPROCESSING:
+            results = _req_pool.map(_get_url_data, req_urls)
+        else:
+            for req_url in req_urls:
+                results.append(_get_url_data(req_url))
     return results
 
 
-def _swap_protocol(url):
-    if 'http://' in url:
-        return url.replace('http://', 'https://')
-    else:
-        return url.replace('https://', 'http://')
+_pool = Pool(processes=3)
 
 
-def _get_urls(urls):
-    additional = []
-    for url in urls:
-        additional.append(_swap_protocol(url))
-    urls.extend(additional)
-    return urls
-
-
-# _pool = Pool(processes=3)
-def _get_counts(urls):
+def _get_counts(urls, count_types):
     counts = {}
-    type_order = []
-    args_list = []
-    for type_, config in COUNT_URLS.items():
-        counts[type_] = 0
-        type_order.append(type_)
-        args_list.append((config, urls))
-
-    for idx, type_count in enumerate(map(_get_urls_data, args_list)):
-        type_ = type_order[idx]
-        for type_count in type_count:
-            counts[type_] += type_count
+    for type_, config in count_types.items():
+        for result in _get_urls_data((config, urls)):
+            if isinstance(result, dict):
+                # manually setting keys on social data here
+                for key, value in result.items():
+                    if key not in counts:
+                        counts[key] = 0
+                    counts[key] += value
+            else:
+                if type_ not in counts:
+                    counts[type_] = 0
+                counts[type_] += result
     return counts
 
 
@@ -154,12 +172,12 @@ def _count_diff(existing, new):
 
 
 @retriable(sync=True)
-def get_social_counts(site, obj, site_url, count=0):
+def get_social_counts(site, obj, site_url, count_types, count=0):
     counts = {}
     site_path = '/'.join(site.getPhysicalPath())
     obj_path = '/'.join(obj.getPhysicalPath())
     rel_path = obj_path[len(site_path):].strip('/')
-    print('Looking up ' + rel_path)
+    print('Looking up ' + obj_path)
 
     urls = [site_url.rstrip('/') + '/' + rel_path]
     registry = getUtility(IRegistry)
@@ -177,8 +195,7 @@ def get_social_counts(site, obj, site_url, count=0):
         rel_path = redirect[len(site_path):].strip('/')
         urls.append(site_url.rstrip('/') + '/' + rel_path)
 
-    urls = _get_urls(urls)
-    counts = _get_counts(urls)
+    counts = _get_counts(urls, count_types)
 
     if not _has_data(counts):
         return
@@ -206,7 +223,20 @@ def retrieve(site):
     registry = getUtility(IRegistry)
     site_url = registry.get('plone.public_url')
     if not site_url:
+        logger.info("No public URL is set; skipping site %s" % site)
         return
+
+    # Which counting strategies should we use?
+    count_types = COUNT_TYPES.copy()
+    site_id = registry.get(MATOMO_SITE_ID, None)
+    base_url = registry.get(MATOMO_BASE_URL, None)
+    token_auth = registry.get(MATOMO_TOKEN_AUTH, None)
+    if not site_id or not base_url or not token_auth:
+        # filter all _matomo count types
+        for ctype in list(count_types.keys()):
+            if ctype.endswith('_matomo'):
+                del count_types[ctype]
+
     catalog = getToolByName(site, 'portal_catalog')
     count = 0
     for brain in catalog(review_state='published'):
@@ -217,7 +247,7 @@ def retrieve(site):
         count += 1
         try:
             obj = brain.getObject()
-            get_social_counts(site, obj, site_url, count)
+            get_social_counts(site, obj, site_url, count_types, count)
             logger.info('retrieved social stats for: %s' % path)
         except Exception:
             logger.warn('error getting social count totals for: %s' % path,