+++ HISTORY.rst
@@ -7,6 +7,13 @@ Changelog
 - Do not error when upgrade hasn't been run for only allow backend urls
   [vangheem]
 
+- Handle incorrectly configured backend url/sheild settings
+  [vangheem]
+
+- Add Mamoto support (CastleCMS API control panel settings, Twitter & Facebook share counting and Analytics display);
+  remove EOL'd Facebook and LinkedIn API social counts
+  [tkimnguyen]
+
 
 2.2.2 (2018-11-26)
 ------------------